fix symbolics, rearrange imports and add some source files
